// autogpt_platform/frontend/src/app/api/__generated__/models/credentialsMetaResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { CredentialsMetaResponseType } from "./credentialsMetaResponseType";
import type { CredentialsMetaResponseTitle } from "./credentialsMetaResponseTitle";
import type { CredentialsMetaResponseScopes } from "./credentialsMetaResponseScopes";
import type { CredentialsMetaResponseUsername } from "./credentialsMetaResponseUsername";
import type { CredentialsMetaResponseHost } from "./credentialsMetaResponseHost";

export interface CredentialsMetaResponse {
  id: string;
  provider: string;
  type: CredentialsMetaResponseType;
  title: CredentialsMetaResponseTitle;
  scopes: CredentialsMetaResponseScopes;
  username: CredentialsMetaResponseUsername;
  /** Host pattern for host-scoped or MCP server URL for MCP credentials */
  host?: CredentialsMetaResponseHost;
}